backend/pyproject.toml
[project]
name = "enttok-backend"
version = "0.1.0"
description = "Enttok-E local backend service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic[email]>=2.0.0",
  "python-dotenv>=1.0.0",
  "google-auth>=2.29.0",
  "google-auth-oauthlib>=1.2.0",
  "google-api-python-client>=2.127.0",
  "httpx>=0.27.0",
  # Memory system dependencies
  "chromadb>=0.4.0",
  "sentence-transformers>=2.2.0",
  "mcp>=1.0.0",
]

